daemon/internal: adjust some build-tags
