--- daemon/internal/platform/platform_unix.go
@@ -1,16 +1,0 @@
-//go:build !windows
-
-package platform
-
-import (
-	"golang.org/x/sys/unix"
-)
-
-// runtimeArchitecture gets the name of the current architecture (x86, x86_64, i86pc, sun4v, ...)
-func runtimeArchitecture() (string, error) {
-	utsname := &unix.Utsname{}
-	if err := unix.Uname(utsname); err != nil {
-		return "", err
-	}
-	return unix.ByteSliceToString(utsname.Machine[:]), nil
-}
